requirements.txt
asgiref==3.2.10
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
ClickUz==0.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
EasyProcess==0.3
entrypoint2==0.2.4
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mss==6.1.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyscreenshot==2.3
pytz==2020.1
pyzmq==19.0.2
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
twilio==6.50.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
